refactor(bus): Move types into typing dir
Moved type Bus into separate typing dir so it isn't in package main
