+++ bus/bus.go
@@ -6,11 +6,11 @@ import (
 	"log"
 	"math/rand"
 	"os"
-	"sync"
 	"time"
 
 	mqtt "github.com/eclipse/paho.mqtt.golang"
 	"github.com/eisandbar/BusPool/bus/path"
+	. "github.com/eisandbar/BusPool/bus/typing"
 	"github.com/golang/geo/s2"
 )
 
@@ -31,18 +31,6 @@ func newBus(id int) *bus {
 		},
 		path.DumbPathFinder{},
 	}
-}
-
-type Bus struct {
-	Id           int
-	Time         time.Time
-	Capacity     int
-	Occupancy    int         // Number of passengers including those not yet picked up
-	Location     s2.LatLng   // Current location
-	Clients      []s2.LatLng // Clients that still need to be picked up
-	Destinations []s2.LatLng // Client drop-off locations
-	Path         []s2.LatLng
-	sync.RWMutex
 }
 
 type Request struct {
